pylong_fromlong + py_decref updates
pylong_fromlong: change pyobject heap alloc size to unknown because in
reality it is greater than just struct pyobject; so if we just set it to
exactly pyobject size then in later stages, may not be in the right
bounds. for now, let's just use
unknown. if necessary, model more accurately later.

py_decref -> py_dealloc: analyzer uses gimple ir to analyze paths. If it
recognizes a function call/macro or something, then it won't go to known
function subclass but instead just simulate the call directly. For this
reason, since py_dealloc is not expanded into the actual code in gimple,
let us analyze that instead. Also added check to see if heap region is
already deallocated for py_decref.

	reason: because analyzer uses gimple ir to analyze paths it sees

foo(){
...
Py_DECREF(obj)
}

void Py_DECREF .....

so 2 calls to Py_DECREF. It will analyze both and for some reason
deallocate both (even tho objects are not supposed to be the same?) For
this reason we must add the check; otherwise will seg fault.