impl stash global variables (for PyTypeObjects) + fix bugs
PyTypeObjects are global variables in CPython. Similar to stashing types
at the translation unit, we do the same for global variables to retrieve
this information. Using this information, we can then build a pointer to
the relevant PyTypeObject when setting ob_type during PyObject
construction (e.g PyList_New). This commit only adds the infrastructure
needed; actually changing implementation of how ob_type is set
(currently it is set to unknown) has not been done yet.

Bugs fixed:
Now that we are accurately modeling the nested struct view during object
construction, the WIP check_pyobj_refcnt is outdated and would throw a
segfault. Temporarily commented this out for now.

Additionally, there was a bug with how we used
get_or_create_unknown_svalue, which expects a tree node representing a
type. Previously we passed in the raw field_decl. We need to pass in
its type (FIELD_DECL::type), not the field_decl itself. So fix is
wrapping the parameter with TREE_TYPE(ob_item_field)