pylist_append: add type checking and null check
CPython performs type checking prior to pylist_append with the macro
PyList_Check(op), which eventually calls the following:

static inline int
PyType_HasFeature(PyTypeObject *type, unsigned long feature)
{
    unsigned long flags;
    // PyTypeObject is opaque in the limited C API
    flags = PyType_GetFlags(type);
    flags = type->tp_flags;
    return ((flags & feature) != 0);
}

This is rather inconvenient to simulate directly, because we would have
to manually include the macro definitions of magic numbers such as
Py_TPFLAGS_LIST_SUBCLASS (in the function above, it would be the feature
argument). For convenience, we can just check to see if the
svalue of the ob_type field of the object that we received as argument
points to the global region for PyList_Type. To be
specific, this approach only applies for PyList_Type specifically and
not subclasses; so it's somewhat of a temporary measure.

Null check, by comparison, is rather straightforward.