implement pylist_append
This commit completes a series which enable a first approximation
analyzer support of PyList_Append. Some concessions were made to
simplify the analysis process when comparing kf_PyList_Append with the
real implementation. In particular, PyList_Append performs some
optimization in list_resize to try and avoid calls to realloc if
possible. For simplicity, we assume that realloc is called every time.
Also, we grow the size by just 1 (to ensure enough space for adding a
new element) rather than the heuristics that the actual implementation
follows.

Additionally, the implementation of kf_PyList_New was changed to support
kf_PyList_Append. Since the analyzer deterministically picks the same
sequence of IDs for heap allocations, when we analyze a program where
multiple appends occur, problems may occur with respect to state
merging. For example, one realloc path may have created a new buffer
(ob_item) and copied all its contents while another path has kept the
same buffer and just resized the buffer. In the former path, the old
buffer will have been poisoned whereas the in the latter path, it is
still relevant. The analyzer may then attempt to merge these paths,
causing all buffers to be kept and their contents to be marked UNKNOWN. We
will then lose any capability for analyzing the contents of the buffers.
A similar situation may occur if we initialize a buffer as NULL in
PyList_New.

For these reasons, the ob_item buffer is now initialized as an
unmergeable svalue. Moreover, if the size given is 0, rather than
initializing it as a NULL, it will be initialized as a heap allocation
of size 0.