Fix: Make  malloc state machine aware of PyObject heap allocations
Sequential heap allocations of PyObjects via PyList_New or
PyLong_FromLong etc caused undesirable state merges between the success
state and the failure state of the heap allocated object before
transition for the next statement. This made reference count checking
(or any other form of checking the property of the PyObject) impossible
in some circumstances. It was caused because the malloc state machine
was not transitioning the pointer of the region from "start" to
"assumed-non-null". To fix this problem, we explicitly transition the
state in a similar way to region_model::on_realloc_with_move.